Update patch set 2
Patch Set 2:

(2 comments)

I have tried to update the API calls I am using per the changes in the peer API source, but it's still not working. The swagger content does not match the code, so I don't know which to use. I have tried all combinations and it's not working.

Patch-set: 2