Update patch set 2
Patch Set 2:

Re "where this is anchored", I meant which data model item is the parent of these items, and has an attribute which is the ID of the items. I assumed, but it would be good to clarify that the RTU items are a list child element of the user.

Patch-set: 2
Label: Code-Review=0