Update patch set 1
Patch Set 1: Code-Review-1

Federation uses the access type code in subscriptions to populate the access type code in imported solution revisions.  If the access type code in subscriptions is going away then the access type code in solution revisions also needs to be removed or there needs to be agreement on some other way to populate the solution revision access type code.

Patch-set: 1