Update patch set 1
Patch Set 1:

Ok that's a fair concern. Access type is definitely going away on revisions, it's just a question of timing because the change is disruptive.  You could use a constant value until then.

Please correct me if I missed it, but I don't see on the Portal admin UI for federation  a way to enter the access type code on a subscription, it appears to be a hidden/secret feature only available via Swagger and/or direct database manipulation.

Patch-set: 1